Remove --enable-optimization
it seems to be gone for some reason. In https://trac.sagemath.org/ticket/17635
they also had some trouble with this and had to remove that flag. I am not sure
what's going on but since it's not listed in the common options on the homepage
of fflas-ffpack, I don't think it's important.